feat: separate getFiles fn from copy script
